Eliminate failure_summary/controller
Moved logic to either the parent controller or into helpers and
the FailureSummaryTab
