+++ ui/plugins/failure_summary/main.html
@@ -1,4 +1,4 @@
-<div ng-controller="BugsPluginCtrl">
+<div>
   <div class="w-100 h-100">
     <failure-summary-tab
       suggestions="suggestions"
